+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 0.9 (unreleased)
 
 - Added `alias` command to map alternate domains to existing models.
+- Updated `Page.html` to return a string.
+- Added `Page.soup` to return parsed HTML using Beautiful Soup.
 
 # 0.8 (2021-05-12)